new heuristics: check for elements with eligible `position` styles
In practice, most cookie notices are implemented with page overlays.
There aren't many ways to accomplish this with CSS, and it seems like
they generally use `position: fixed` or `position: sticky`.

Doing this as the first step drastically filters down the HTML search
space for further textual inspection, which can be more involved.

Not all sites implement cookie banners this way, but this check is
effective enough that it's worth using as a primary approach. Additional
heuristics can be added alongside it in the future.