improve scroll blocking detection
By checking for elements that create stacking contexts, the scroll
blocking detection can be improved to filter out false positives. In
particular, cookiemonster can declare that if the front-most element in
the page takes up the full page and contains no textual content, it's
highly likely to be a broken overlay element.